requirements.txt
ultralytics
Pillow
opencv-python
gradio
pandas